requirements.txt
Django==4.0.2
djangorestframework
markdown
django-filter
firebase_admin
